refactored, improved naming and added documentation
