+++ src/client/components/Utility/Head.js
@@ -1,6 +1,7 @@
 import React from 'react'
 
-const Head = ({ title = "", description = "", stylePath = "", metas = [], children }) => (
+
+const Head = ({ title = "", description = "", metas = [], children }) => (
 	<>
 		{title && <title>{title}</title>}
 		{description && <meta name="description" content={description} />}
@@ -11,7 +12,6 @@ const Head = ({ title = "", description = "", stylePath = "", metas = [], childr
 		<meta charSet="UTF-8" />
 		<meta name="viewport" content="width=device-width, initial-scale=1, viewport-fit=cover" />
 		<link href="https://fonts.googleapis.com/css2?family=Rubik&display=swap" rel="stylesheet" />
-		<link rel="stylesheet" href={stylePath} />
 		{children}
 	</>
 )
