+++ src/client/pages/landing/index.js
@@ -1,6 +1,7 @@
 import renderWithCache from '../../utils/ssrUtils/renderWithCache'
 import App from './App'
 
+
 renderWithCache(App)
 
 if (module.hot) {
