+++ src/server/development.js
@@ -10,19 +10,18 @@ const webpackHMR = require('webpack-hot-middleware')
 const webpackDevConfig = require('../config/webpack.config.dev.js')
 const app = require('./server')
 
+
 const PORT = process.env.PORT || 3000
 
-
+// configure webpack
 const compiler = webpack(webpackDevConfig)
-
+// add dev middleware to our custom server
 app.use(webpackDevMiddleware(compiler, {
 	stats: { colors: true },
 	publicPath: webpackDevConfig.output.publicPath
 }))
-
+// add hot module reload middleware
 app.use(webpackHMR(compiler, { heartbeat: 2000 }))
 
+// you know... listen.
 app.listen(PORT, () => console.log(`Serving on port ${PORT}`))
-
-
-
